requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1

# Core dependencies
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.3.0

# Face recognition stack
insightface==0.7.3
onnxruntime==1.19.0       # FIXED for Python 3.12
faiss-cpu==1.13.0         # FIXED for Python 3.12

# Async server utils
motor==3.6.0
python-multipart==0.0.9
aiofiles==23.2.1

aiortc==1.6.0
av==12.0.0